[language] Clean up genesis generation and move it under language/tools
This moves the genesis generation into `language/tools` which is a better home for the tool.
It also cleans it up by refactoring the code and removing the usage of `TransactionExecutor`.
`TransactionExecutor` is removed and usage of it are using the `VMRuntime` directly. That is
in line with the work that @runtian-zhou was planning to do.
I verified the genesis generated is identical to what it used to be.
`vm-genesis/src/lib.rs` shows as added and removed which is fine as the one re-worked is
better read with no diff in place and it should be very straightforward
