+++ consensus/Cargo.toml
@@ -61,7 +61,7 @@ proptest = "0.9.4"
 rusty-fork = "0.2.2"
 tempfile = "3.1.0"
 
-vm-genesis = { path = "../language/vm/vm-genesis", version = "0.1.0" }
+vm-genesis = { path = "../language/tools/vm-genesis", version = "0.1.0" }
 vm-validator = { path = "../vm-validator", version = "0.1.0" }
 
 [features]
