+++ executor/Cargo.toml
@@ -42,7 +42,7 @@ rusty-fork = "0.2.1"
 grpc-helpers = { path = "../common/grpc-helpers", version = "0.1.0" }
 lcs = { path = "../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 storage-service = { path = "../storage/storage-service", version = "0.1.0" }
-vm-genesis = { path = "../language/vm/vm-genesis", version = "0.1.0" }
+vm-genesis = { path = "../language/tools/vm-genesis", version = "0.1.0" }
 transaction-builder = { path = "../language/transaction-builder", version = "0.1.0" }
 
 [features]
