+++ language/e2e-tests/Cargo.toml
@@ -22,7 +22,7 @@ libra-types = { path = "../../types", version = "0.1.0", features = ["fuzzing"] 
 transaction-builder = { path = "../transaction-builder", version = "0.1.0", features = ["fuzzing"]}
 vm = { path = "../vm", version = "0.1.0" }
 vm-cache-map = { path = "../vm/vm-runtime/vm-cache-map", version = "0.1.0" }
-vm-genesis = { path = "../vm/vm-genesis", version = "0.1.0" }
+vm-genesis = { path = "../tools/vm-genesis", version = "0.1.0" }
 vm-runtime = { path = "../vm/vm-runtime", version = "0.1.0" }
 vm-runtime-types = { path = "../vm/vm-runtime/vm-runtime-types", version = "0.1.0" }
 proptest = "0.9.3"
