+++ language/tools/test-generation/src/lib.rs
@@ -35,9 +35,7 @@ use vm::{
 };
 use vm_cache_map::Arena;
 use vm_runtime::chain_state::TransactionExecutionContext;
-use vm_runtime::{
-    data_cache::BlockDataCache, runtime::VMRuntime, txn_executor::TransactionExecutor,
-};
+use vm_runtime::{data_cache::BlockDataCache, runtime::VMRuntime};
 use vm_runtime_types::value::Value;
 
 /// This function calls the Bytecode verifier to test it
@@ -98,9 +96,17 @@ fn execute_function_in_module(
         let data_cache = BlockDataCache::new(state_view);
         let context = TransactionExecutionContext::new(*MAXIMUM_NUMBER_OF_GAS_UNITS, &data_cache);
         let gas_schedule = runtime.load_gas_schedule(&context, &data_cache)?;
-        let mut txn_executor =
-            TransactionExecutor::new(&gas_schedule, &data_cache, TransactionMetadata::default());
-        txn_executor.execute_function(&runtime, &module_id, &entry_name, args)
+        let txn_data = TransactionMetadata::default();
+        let mut interpreter_context =
+            TransactionExecutionContext::new(txn_data.max_gas_amount(), &data_cache);
+        runtime.execute_function(
+            &mut interpreter_context,
+            &txn_data,
+            &gas_schedule,
+            &module_id,
+            &entry_name,
+            args,
+        )
     }
 }
