+++ language/tools/vm-genesis/Cargo.toml
@@ -11,20 +11,22 @@ publish = false
 
 [dependencies]
 anyhow = "1.0"
+lazy_static = "1.3.0"
+rand = "0.6.5"
+
+bytecode-verifier = { path = "../../bytecode-verifier", version = "0.1.0" }
 lcs = { path = "../../../common/lcs", version = "0.1.0", package = "libra-canonical-serialization" }
 libra-config = { path = "../../../config", version = "0.1.0" }
-transaction-builder = { path = "../../transaction-builder", version = "0.1.0"}
 libra-crypto = { path = "../../../crypto/crypto", version = "0.1.0" }
-stdlib = { path = "../../stdlib", version = "0.1.0" }
 libra-state-view = { path = "../../../storage/state-view", version = "0.1.0" }
 libra-types = { path = "../../../types", version = "0.1.0" }
 parity-multiaddr = "0.6.0"
-vm = { path = "../", version = "0.1.0" }
-vm-cache-map = { path = "../vm-runtime/vm-cache-map", version = "0.1.0" }
-vm-runtime = { path = "../vm-runtime", version = "0.1.0" }
-vm-runtime-types = { path = "../vm-runtime/vm-runtime-types", version = "0.1.0" }
-lazy_static = "1.3.0"
-rand = "0.6.5"
+stdlib = { path = "../../stdlib", version = "0.1.0" }
+transaction-builder = { path = "../../transaction-builder", version = "0.1.0"}
+vm = { path = "../../vm", version = "0.1.0" }
+vm-cache-map = { path = "../../vm/vm-runtime/vm-cache-map", version = "0.1.0" }
+vm-runtime = { path = "../../vm/vm-runtime", version = "0.1.0" }
+vm-runtime-types = { path = "../../vm/vm-runtime/vm-runtime-types", version = "0.1.0" }
 
 [dev-dependencies]
 proptest = "0.9.3"
