+++ language/tools/vm-genesis/genesis/genesis.blob
