+++ language/vm/README.md
@@ -52,6 +52,5 @@ simple abstraction over the file format. Additionally, a set of
 ├── cost-synthesis  # Infrastructure for gas cost synthesis
 ├── src             # VM core files
 ├── tests           # Proptests
-├── vm-genesis      # Helpers to generate a genesis block, the initial state of the blockchain
 └── vm-runtime      # Interpreter and runtime data types (see README in that folder)
 ```
