+++ state-synchronizer/Cargo.toml
@@ -37,7 +37,7 @@ bytes = "0.5"
 config-builder = { path = "../config/config-builder", version = "0.1.0" }
 libra-crypto = { path = "../crypto/crypto", version = "0.1.0" }
 parity-multiaddr = "0.6.0"
-vm-genesis = { path = "../language/vm/vm-genesis", version = "0.1.0" }
+vm-genesis = { path = "../language/tools/vm-genesis", version = "0.1.0" }
 transaction-builder = { path = "../language/transaction-builder", version = "0.1.0" }
 channel = { path = "../common/channel", version = "0.1.0" }
